CKMultipartParser: implement a simple MIME multipart parser.
The parser reads the content of an NSData and split it into multiple
chunks. The chunks are not parsed more than necessary; the headers part
is parsed into a simple dictionary with key/values, but the values are
not parsed (eg. Content-Disposition).

With the parser, there is also an implementation of the KMP Pattern
Matching algorithm in an NSData extension. This function is used
intensively to look for the "boundary" pattern in the data buffer.