CKImageView: improves the performances by
(1) replacing the custom rendering done in -drawRect by delegating to a
UIImageView.
(2) removing any image resizing (via CKImageLoader); only use the
original image, it's up to the client to give the CKImageView with
images of the right size or doing the resizing somewhere else.
(3) does not support -aspectFill anymore, it's delegated to the
UIImageView.