CKWebRequest2: supports application/x-www-form-urlencoded
Configures the request for send a specified dictionary as a POST request
to the HTTP server, in the same way that a browser does when a user has
filled in an HTML form and presses the submit button. This will cause
the request to pass the data to the server using the content-type
application/x-www-form-urlencoded.