CKNetworkActivityManager: fixes an issue with multithreading.
There was a couple of issue with this object,

(1) Multithreading: since the NSMutableSet is not thread-safe, we now
use a performSelectorOnMainThread:withObject:waitUntilDone: to ensure
that only one thread access it. We use the main thread as the central
access point to avoid thread for blocking (we don't use mutexes).

(2) The NSMutableSet is used instead of a NSMutableArray. Search
throught a NSMutableSet is faster O(1) instead of O(n^2).