CKCoreDataManager: fixes issue with migration by supporting explicit model URL.
Core Data was initialized by automatically merging models found in the
bundle. But merging will fail if different version of models are found.
It happened with the CKStore that now has 2 versions to support.

The solution was to explicitely gives the URL of the momd folder
instead of using the merging.

I had to change the interface to expose this, and I took the opportunity
to remove the static singleton from the CKCoreDataManager--because a
"shared" manager does not make any sense.

So, the singleton is now at the CKStore level, where all instances share
the same CKCoreDataManager, which make sense.

BUGS:

There is also an issue issue that should be addresses at some point,
such as the default filename for the database (which is based on the
kCFBundleNameKey. If the bundle name *changes*, the database won't be
found, a new one will be created and the user would have lost its
changes.

For it to work, the "default" naming convention should be based *at
least* on the bundle identifier.