CKWebRequest2: supports authentication.
Adds a property for using a default NSURLCredential with this request.
Implements the authentication challenge by checking is there is a
provided credential--if not, check the NSURLCredentialStorage for a
credential.

This implements also adds a property "validatesSecureCertificate" to be
able to ignore self-signed certificates. The default is to NO (for
compatibility reason) but should definitely be set to YES.