Fix bindings.
Sometimes, when unregistering a binding due to a weak ref deallocation, we called the unregister binding method of the binding manager.
During this method, we reset the binding that potentially release a block. By releasing this block, some objects that have been retained by the block could be deallocated and catches by another binding weak ref or a binding context weak ref. This one will also and synchronously unregister itself or clear the same context that is being modified by the first binding. That resulted in unexpected results and crashs.