Adds CKResourceDependencyContext.
This is used when applying stylesheets on view and cell controllers. By encapsulating the style application with a dependency context, the conversion system automatically knows that he must populate the file path set for resources that are loaded while this context is opened (Images, …).
Afterward, we can forward this dependency list to the style manager that can register for updates and reload appropriately when one or more dependencies gets updated.