Removes ColorPalette. Implements Value injection in CascadingTree.
CascadingTree now provides two ways to organize the data : Inheritance and Injection

Inheritance allow multiple inheritance and overriding of key value sets and hierarchy.
Injection allow to replace a value for a specific key by referencing true value defined elsewhere in the cascading tree.

Here is an example of data injection :

{
     "ColorPalette" : {
         "MyBackgroundColor" : "redColor"
     },

     "UIViewController" : {
         "view" : {
                "backgroundColor" : "ColorPalette. MyBackgroundColor"
          }
     }
}

In this example, the background colour of the view in UIView controller will get the value of ColorPalette.MyBackgroundColor injected and will be red as a result.

Injection could be very useful to group colours or metrics for the UI. By grouping this values in dictionaries we can externalize a whole set of attributes for the UI in specific files that we'll import where needed.
Another advantage is that we can wrap it in @ipad or @iphone and by this way defines metrics or colours that are specific for a device and use the same key path to inject the values in your layouts and stylesheets.