CKCascadingTree : Allows per os Version as well as per device key value definition. This is available in all systems inheriting cascading tee like stylesheets, mappings, object graphs, mock, ...
Here is an extreme sample of what it allows. We can specify different property/values for the targeted "view" depending on the os version and the device model.

Supported operators for device version are "<", "<=", ">", ">=", "=", "==", "!=".
The format for selecting property/values per os version is : "@ios whitespace operator whitespace version"

       "view" : {
            "@ios == 6.1" : {
                "@iphone5" : {
                    "backgroundColor" : "redColor"
                },
                "@iphone" : {
                    "backgroundColor" : "greenColor"
                },
                "@ipad" : {

                }
            },
            "@ios < 6.1" : {
                "@iphone5" : {
                    "backgroundColor" : "yellowColor"
                },
                "@ipad" : {
                    "backgroundColor" : "purpleColor"
                }
            },
            "@ios >= 7" : {
                "@ipad" : {
                    "backgroundColor" : "blueColor"
                },
                "@iphone" : {
                    "backgroundColor" : "darkGrayColor"
                }
            }
        }

Device selectors are : "@iphone", "@iphone5" and "@ipad"

if defining only @iphone, all the iPhone models will be targeted.
if defining only @iphone5, only the phones with aspect ratio of the iphone5 will be targeted.
if defining only @ipad, only the iPad devices will get targeted.