[iOS7] UIBarButtonItem + StyleSheet : Manages alignmentRectInsets for UIBarButtonItem customized using stylesheets.
Customizing barButtonItems in stylesheet internally creates a button and set it as custom view in the UIBarButtonItem. On iOS7, these buttons were not aligned correctly in the navigation bar due to the auto layout constraints.

We now manages custom alignmentRectInsets on these buttons by getting the parent navigation bar and checking if the corresponding bar button item is the left or right bar button item of the top navigation item.