CKObject : Adds support for CKCollection property changes informal protocol.
By defining an instance method propertyNameObjectsAdded:indexes: or propertyNameObjectsRemoved:indexes: on a class inheriting CKObject, the instance will automatically register/unregister for watching changes in the collection with the specified property name and call the implemented selector when changes occurs.

This is particularly useful when other properties for this instance must be updated when the property change.

for example :

@interface MyClass : CKObject
@property (nonatomic,retain) CKArrayCollection* myCollection;
@end

@implementation MyClass

- (void) myCollectionObjectsAdded:(NSArray*)objects indexes:(NSIndexSet*)indexes{
	//Updates other properties using the added objects in myCollection
}

- (void) myCollectionObjectsRemoved:(NSArray*)objects indexes:(NSIndexSet*)indexes{
	//Updates other properties using the removed objects from myCollection
}

@end