Fixes potential Crash with tableview and collection views.
When computing size if cells have custom layout, sometimes, bindings on the cell context can act on a weak cellController's ref. But as we need to compute size without the cell itself, we crate and reuse cells in cache to init and setup them then compute the size. Bindings weren't cleared on the cell context afterward and could lead to crashes in this particular case.