CKCascadingTree: Implements an alias and a transformer system.
* Alias

Keys and string values in the cascading tree can be replaced when loading the tree and being replaced by aliases that are registered in the Cascading tree using the following method:

+ (void)registerAlias:(NSString*)alias forKey:(NSString*)key;

For Example, CKLayoutBox defines an alias that allows to do:

"layout" : [ <LayoutBoxes> ] instead of "layoutBoxes" : [ <LayoutBoxes> ]

Both are valid.

* Transformers

Transformer block can be registered in the Cascading tree with a predicate block. If the predicate returns yes, the transformer block will be executed and allows to modify the structure of the cascading tree using the values and the keys that are available after parsing the JSON. This allows to simplify the syntax for some specific use cases.

For example in CKLayoutBox, we register 2 transformers for dictionnaries defining UIView or CKLayoutBox objects:

1. { "<ClassName>" : "<Name>" } will be replaced by { "@class" : "<ClassName>", "name" : "<Name>" }

2. { "<ClassName>" : [ <LayoutBoxes> ] } will be replaced by { "@class" : "<ClassName>", "layoutBoxes" : [ <LayoutBoxes> ] }